rename(notice): rename departmentnotice -> notice, notice -> noticeDetails
