+++ src/main/java/com/dongsoop/dongsoop/notice/entity/Notice.java
@@ -1,10 +1,16 @@
 package com.dongsoop.dongsoop.notice.entity;
 
+import com.dongsoop.dongsoop.department.DepartmentType;
+import jakarta.persistence.CascadeType;
+import jakarta.persistence.Embeddable;
 import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
 import jakarta.persistence.Id;
-import java.time.LocalDate;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import java.util.Objects;
 import lombok.AllArgsConstructor;
-import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 @Entity
@@ -13,14 +19,35 @@ import lombok.NoArgsConstructor;
 public class Notice {
 
     @Id
-    @Getter
-    private Long id;
+    private DepartmentNoticeKey id;
 
-    private String writer;
+    @Embeddable
+    @AllArgsConstructor
+    @NoArgsConstructor
+    public static class DepartmentNoticeKey {
+        @Enumerated(EnumType.STRING)
+        private DepartmentType type;
 
-    private String title;
+        @ManyToOne(cascade = CascadeType.PERSIST)
+        @JoinColumn
+        private NoticeDetails noticeDetails;
 
-    private String link;
+        // JPA 엔티티 비교 및 캐싱 시 사용된다
+        @Override
+        public boolean equals(Object o) {
+            if (this == o) {
+                return true;
+            }
+            if (o == null || getClass() != o.getClass()) {
+                return false;
+            }
+            DepartmentNoticeKey that = (DepartmentNoticeKey) o;
+            return type == that.type && Objects.equals(noticeDetails, that.noticeDetails);
+        }
 
-    private LocalDate createdAt;
+        @Override
+        public int hashCode() {
+            return Objects.hash(type, noticeDetails);
+        }
+    }
 }
