+++ src/main/java/com/dongsoop/dongsoop/notice/repository/NoticeDetailsRepository.java
@@ -1,7 +1,7 @@
 package com.dongsoop.dongsoop.notice.repository;
 
-import com.dongsoop.dongsoop.notice.entity.Notice;
+import com.dongsoop.dongsoop.notice.entity.NoticeDetails;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NoticeDetailsRepository extends JpaRepository<Notice, Long> {
+public interface NoticeDetailsRepository extends JpaRepository<NoticeDetails, Long> {
 }
